% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{BINOMIAL}
\alias{BINOMIAL}
\alias{new_BINOMIAL}
\title{Factory for a BINOMIAL distribution object}
\usage{
new_BINOMIAL(p_size, p_prob)
}
\arguments{
\item{p_size}{number of trials}

\item{p_prob}{probabilty of success of each trial}
}
\value{
An object of class \code{\link{DISTRIBUTION}}, \code{BINOMIAL}
}
\description{
Returns a BINOMIAL distribution object that produce random numbers
from a binomial distribution using the \code{\link{rbinom}} funtion
}
\examples{
myDistr <- new_BINOMIAL(0,1)
myDistr$rfunc(10)
}
\author{
John Aponte
}
