% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{cinqnum.DISTRIBUTION}
\alias{cinqnum.DISTRIBUTION}
\title{Generic function for a distribution}
\usage{
\method{cinqnum}{DISTRIBUTION}(x, n)
}
\arguments{
\item{x}{an object of  class \code{\link{DISTRIBUTION}}}

\item{n}{number of dranws}
}
\value{
a list with the  mean, sd, 95%upper and lower centiles and median
}
\description{
Generantes n random numbers from the distribution, using the seed of the
object, so always return the same value. Internal function to be used in the
summary
}
