% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{cinqnum.NA}
\alias{cinqnum.NA}
\title{And optimized version for NA distribution}
\usage{
\method{cinqnum}{NA}(x, n)
}
\arguments{
\item{x}{an object of  class \code{\link{DISTRIBUTION}}}

\item{n}{number of dranws}
}
\value{
a list of NA
}
\description{
And optimized version for NA distribution
}
