% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{cinqnum}
\alias{cinqnum}
\title{cinqnum}
\usage{
cinqnum(x, ...)
}
\arguments{
\item{x}{an object of class \code{\link{DISTRIBUTION}}}

\item{...}{further parameters}
}
\description{
Make a list with 5 numbers of the distribution (mean_, sd_, lci_, uci_, median_).
}
\details{
Uses the stored seed to have the same drawns always and produce the same numbers
This is an internal function for the summary function
}
