% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convdistr-package.R
\docType{package}
\name{convdistr-package}
\alias{convdistr}
\alias{convdistr-package}
\title{convdistr: Convolute Probabilistic Distributions}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Convolute probabilistic distributions using the random generator function of each distribution. A new random number generator function is created that perform the mathematical operation on the individual random samples from the random generator function of each distribution. See the documentation for examples.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/johnaponte/convdistr}
  \item \url{https://johnaponte.github.io/convdistr/}
}

}
\author{
\strong{Maintainer}: Aponte John \email{john.j.aponte@gmail.com} (\href{https://orcid.org/0000-0002-3014-3673}{ORCID})

}
\keyword{internal}
