% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convdistr.R
\docType{package}
\name{convdistr}
\alias{convdistr}
\alias{convdistr-package}
\title{convdistr: A package usefull for convolution of distributions.}
\description{
The convdistr provides tools to definde distribution objects and make 
mathematical operations with them. It keep track of the results as if they
where scalar numbers but mantain the ability to obtain randoms samples of
the convoluted distributions.
}
