% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ggDISTRIBUTION}
\alias{ggDISTRIBUTION}
\title{Plot of \code{\link{DISTRIBUTION}} objects using \code{\link[ggplot2]{ggplot2}}}
\usage{
ggDISTRIBUTION(distribution, n = 10000)
}
\arguments{
\item{distribution}{an object of class \code{\link{DISTRIBUTION}}}

\item{n}{number of drawns}
}
\value{
a \code{\link[ggplot2]{ggplot}} object with the density of the distribution
}
\description{
Plot of \code{\link{DISTRIBUTION}} objects using \code{\link[ggplot2]{ggplot2}}
}
\examples{
x <- new_NORMAL(0,1)
ggDISTRIBUTION(x)
y <- new_DIRICHLET(c(10,20,70))
ggDISTRIBUTION(x)
}
