% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{LOGNORMAL}
\alias{LOGNORMAL}
\alias{new_LOGNORMAL}
\title{Factory for a LOGNORMAL distribution object}
\usage{
new_LOGNORMAL(p_meanlog, p_sdlog)
}
\arguments{
\item{p_meanlog}{mean of the distribution on th elog scale}

\item{p_sdlog}{A numeric that represents the standard deviation on the log scale}
}
\value{
An object of class \code{\link{DISTRIBUTION}}, \code{LOGNORMAL}
}
\description{
Returns a LOGNORMAL distribution object that produce random numbers
from a log normal distribution using the \code{\link{rlnorm}} funtion
}
\examples{
myDistr <- new_LOGNORMAL(0,1)
myDistr$rfunc(10)
}
\author{
John Aponte
}
