% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{metadata}
\alias{metadata}
\alias{metadata.DISTRIBUTION}
\alias{metadata.default}
\title{Metadata for a DISTRIBUTION}
\usage{
metadata(x)

\method{metadata}{DISTRIBUTION}(x)

\method{metadata}{default}(x)
}
\arguments{
\item{x}{a \code{\link{DISTRIBUTION}} object}
}
\value{
A \code{\link{data.frame}} with the metadata of the distributions
}
\description{
Shows the distribution and the oval values of a \code{\link{DISTRIBUTION}}
object
}
\section{Methods (by class)}{
\itemize{
\item \code{DISTRIBUTION}: Metadata for DISTIBUTION objects

\item \code{default}: Metadata for other objects
}}

\note{
The number of columns depends on the dimensions of the distribution.
There will be one column \code{distribution} with the name of the distribution
and one column for each dimension with the names from the \code{oval} field.
}
