% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{NA_DISTRIBUTION}
\alias{NA_DISTRIBUTION}
\alias{new_NA}
\title{Factory for a NA distribution object}
\usage{
new_NA(p_dimnames = "rvar")
}
\arguments{
\item{p_dimnames}{A character that represents the the names of the
dimensions. By default only one dimension with name \code{rvar}}
}
\value{
An object of class \code{DISTRIBUTION}, \code{NA}
}
\description{
Returns an NA distribution object that always return \code{NA_real_}
This is usefull to handle \code{NA}.
By default only one dimension \code{rvar} is produced, but if several
names are provided more columns will be added to the return matrix
}
\examples{
myDistr <- new_NA(p_dimnames = "rvar")
myDistr$rfunc(10)
}
\author{
John Aponte
}
