% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{new_BETA}
\alias{new_BETA}
\title{Factory for a BETA distribution object}
\usage{
new_BETA(p_shape1, p_shape2)
}
\arguments{
\item{p_shape1}{A numeric that represents the lower limit}

\item{p_shape2}{A numeric that represents the upper limit}
}
\value{
An object of class \code{DISTRIBUTION}, \code{BETA}
}
\description{
Returns an BETA distribution object that produce random numbers
from a  beta distribution using the \code{\link{rbeta}} funtion
}
\examples{
myDistr <- new_BETA(1,1)
myDistr$rfunc(10)
}
\author{
John Aponte
}
