% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolution_assoc.R
\name{new_CONVOLUTION_assoc}
\alias{new_CONVOLUTION_assoc}
\title{Convolution with asociation of dimensions}
\usage{
new_CONVOLUTION_assoc(dist1, dist2, op)
}
\arguments{
\item{dist1}{an object of class \code{\link{DISTRIBUTION}}}

\item{dist2}{and object of class \code{\link{DISTRIBUTION}}}

\item{op}{one of `+`,`-`,`*`,`/`}
}
\value{
an object of class \code{\link{DISTRIBUTION}}
}
\description{
In case of different dimnesions of the distribution
this function perform the operation on the common distributions and
add without modifications the other dimensions of the distribution.
}
\details{
If distribution A have dimensions a and b and distribution B have dimensions
b and c, the A + B would produce a distribution with dimensions
a, c,  b+b,
}
