% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolution_comb.R
\name{new_CONVOLUTION_comb}
\alias{new_CONVOLUTION_comb}
\title{Convolution with combination of dimensions}
\usage{
new_CONVOLUTION_comb(dist1, dist2, op, p_dimnames)
}
\arguments{
\item{dist1}{an object of class \code{\link{DISTRIBUTION}}}

\item{dist2}{and object of class \code{\link{DISTRIBUTION}}}

\item{op}{one of `+`,`-`,`*`,`/`}

\item{p_dimnames}{a character vector with the name of the dimensions.
If missing the combination of the individual dimensions will be used}
}
\value{
an object of class \code{\link{DISTRIBUTION}}
}
\description{
In case of different dimnesions of the distribution
this function perform the operation on the combination of the distributions
of both distribution.
}
\details{
If distribution A have dimensions a and b and distribution B have dimensions
b and c, the A + B would produce a distribution with dimensions
a_b,a_c,b_b, b_c
}
\note{
In case of the same dimensions, only the first combination is taken
}
