% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{new_DIRAC}
\alias{new_DIRAC}
\title{Factory for a DIRAC distribution object}
\usage{
new_DIRAC(p_value)
}
\arguments{
\item{p_value}{A numeric that set the value for the distribution}
}
\value{
An object of class \code{DISTRIBUTION}, \code{DIRAC}
}
\description{
Returns an DIRAC distribution object that always return the same number.
}
\examples{
myDistr <- new_DIRAC(p_value = 1)
myDistr$rfunc(10)
}
\author{
John Aponte
}
