% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{new_EXPONENTIAL}
\alias{new_EXPONENTIAL}
\title{Factory for a EXPONENTIAL distribution using confidence intervals}
\usage{
new_EXPONENTIAL(p_rate)
}
\arguments{
\item{p_rate}{A numeric that represents the rate of events}
}
\value{
An object of class \code{DISTRIBUTION}, \code{EXPONENTIAL}
}
\description{
Returns an EXPONENTIAL distribution object that produce random numbers
from an exponential distribution using the \code{\link{rexp}} function
}
\examples{
myDistr <- new_EXPONENTIAL(5)
myDistr$rfunc(10)
}
\author{
John Aponte
}
