% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolution.R
\name{new_MIXTURE}
\alias{new_MIXTURE}
\title{Mixture of \code{\link{DISTRIBUTION}} objects}
\usage{
new_MIXTURE(listdistr, mixture)
}
\arguments{
\item{listdistr}{a list of \code{\link{DISTRIBUTION}} objects}

\item{mixture}{a vector of probabilities to mixture the distributions. Must add 1
If missing the drawns are obtained from the distributions with the same probability}
}
\value{
an object of class \code{MIXTURE}, \code{\link{DISTRIBUTION}}
}
\description{
Produce a new distribution that obtain random drawns of the mixture
of the \code{\link{DISTRIBUTION}} objects
}
\examples{
x1 <- new_NORMAL(0,1)
x2 <- new_NORMAL(4,1)
x3 <- new_NORMAL(6,1)
new_MIXTURE(list(x1,x2,x3))
}
\author{
John J. Aponte
}
