% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{new_POISSON}
\alias{new_POISSON}
\title{Factory for a POISSON distribution using confidence intervals}
\usage{
new_POISSON(p_lambda)
}
\arguments{
\item{p_lambda}{A numeric that represents the expected number of events}
}
\value{
An object of class \code{DISTRIBUTION}, \code{POISSON}
}
\description{
Returns an POISSON distribution object that produce random numbers
from a Poisson distribution using the \code{\link{rpois}} function
}
\examples{
myDistr <- new_POISSON(5)
myDistr$rfunc(10)
}
\author{
John Aponte
}
