% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{NORMAL}
\alias{NORMAL}
\alias{new_NORMAL}
\title{Factory for a NORMAL distribution object}
\usage{
new_NORMAL(p_mean, p_sd, p_dimnames = "rvar")
}
\arguments{
\item{p_mean}{A numeric that represents the mean value}

\item{p_sd}{A numeric that represents the standard deviation}

\item{p_dimnames}{A character that represents the name of the dimension}
}
\value{
An object of class \code{\link{DISTRIBUTION}}, \code{NORMAL}
}
\description{
Returns a NORMAL distribution object that produce random numbers
from a normal distribution using the \code{\link{rnorm}} funtion
}
\examples{
myDistr <- new_NORMAL(0,1)
myDistr$rfunc(10)
}
\author{
John Aponte
}
