% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{omit_NA}
\alias{omit_NA}
\title{Omit \code{NA} distributions from a list of distributions}
\usage{
omit_NA(listdistr)
}
\arguments{
\item{listdistr}{a list of \code{\link{DISTRIBUTION}} objects}
}
\value{
the list without the \code{\link{NA_DISTRIBUTION}}
}
\description{
Omit \code{NA} distributions from a list of distributions
}
