% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.DISTRIBUTION}
\alias{plot.DISTRIBUTION}
\title{plot of  \code{\link{DISTRIBUTION}} objects}
\usage{
\method{plot}{DISTRIBUTION}(x, n = 10000, ...)
}
\arguments{
\item{x}{an object of class \code{\link{DISTRIBUTION}}}

\item{n}{number of observations}

\item{...}{other parameters to the \code{\link{hist}} function}
}
\description{
Plot an histogram of the density of the distribution using random
numbers from the distribution
}
\examples{
x <- new_NORMAL(0,1)
plot(x)
y <- new_DIRICHLET(c(10,20,70))
plot(x) 
}
