% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{rfunc.default}
\alias{rfunc.default}
\title{Default function}
\usage{
\method{rfunc}{default}(x, n)
}
\arguments{
\item{x}{an object of class different from \code{\link{DISTRIBUTION}}}

\item{n}{the number of random samples}
}
\description{
Default function
}
