% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{rfunc.DISTRIBUTION}
\alias{rfunc.DISTRIBUTION}
\title{Generic function for a \code{\link{DISTRIBUTION}} object}
\usage{
\method{rfunc}{DISTRIBUTION}(x, n)
}
\arguments{
\item{x}{an object of class \code{\link{DISTRIBUTION}}}

\item{n}{the number of random samples}
}
\description{
Generic function for a \code{\link{DISTRIBUTION}} object
}
