% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{rfunc}
\alias{rfunc}
\title{Generate random drawns from a \code{\link{DISTRIBUTION}} object}
\usage{
rfunc(x, n)
}
\arguments{
\item{x}{an object}

\item{n}{the number of random samples}
}
\value{
a matrix with as many rows as \code{n} and as many columns as
dimensions have \code{distribution}
}
\description{
This is a generic method that calls the \code{rfunc} slot of the object
}
