% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{same_dimensions}
\alias{same_dimensions}
\title{Check the dimensions of a list of distributions}
\usage{
same_dimensions(listdistr)
}
\arguments{
\item{listdistr}{a list of \code{\link{DISTRIBUTION}} objects}
}
\value{
return \code{TRUE} if all the dimensions are the same
}
\description{
Check the dimensions of a list of distributions
}
