% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{set_seed}
\alias{set_seed}
\title{Modify a the seed of a Distribution object}
\usage{
set_seed(distribution, seed)
}
\arguments{
\item{distribution}{a \code{\link{DISTRIBUTION}} object}

\item{seed}{the new seed}
}
\value{
a code{\link{DISTRIBUTION}} object of the same class
}
\description{
This create a new \code{\link{DISTRIBUTION}} object but with the
specified seed
}
\author{
John J. Aponte
}
