% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.DISTRIBUTION}
\alias{summary.DISTRIBUTION}
\title{Summary of Distributions}
\usage{
\method{summary}{DISTRIBUTION}(object, n = 10000, ...)
}
\arguments{
\item{object}{object of class \code{\link{DISTRIBUTION}}}

\item{n}{the number of random drawns from the distribution}

\item{...}{other parameteres. Not used}
}
\value{
A \code{\link{data.frame}} with as many rows as dimensions had the
distribution and  with the following columns
\itemize{
  \item distribution name
  \item varname name of the dimension
  \item oval value
  \item nsample number of random drawns
  \item mean_ mean value of the sample
  \item sd_ standard deviation of the sample
  \item lci_ lower 95% centile of the sample
  \item median_ median value of the sample
  \item uci_ upper 95% centile of the sample
}
}
\description{
Summary of Distributions
}
\note{
The sample uses the seed saved in the object those it will
provide the same values fir an \code{n} value
}
