% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{UNIFORM}
\alias{UNIFORM}
\alias{new_UNIFORM}
\title{Factory for a UNIFORM distribution object}
\usage{
new_UNIFORM(p_min, p_max, p_dimnames = "rvar")
}
\arguments{
\item{p_min}{A numeric that represents the lower limit}

\item{p_max}{A numeric that represents the upper limit}

\item{p_dimnames}{A character that represents the name of the dimension}
}
\value{
An object of class \code{DISTRIBUTION}, \code{UNIFORM}
}
\description{
Returns an UNIFORM distribution object that produce random numbers
from a  uniform distribution using the \code{\link{runif}} function
}
\examples{
myDistr <- new_UNIFORM(0,1)
myDistr$rfunc(10)
}
\author{
John J. Aponte
}
