% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolution_comb.R
\name{add_total}
\alias{add_total}
\title{Adds a total dimension}
\usage{
add_total(p_distribution, p_totalname = "TOTAL")
}
\arguments{
\item{p_distribution}{an object of class \code{\link{DISTRIBUTION}}}

\item{p_totalname}{the name of the new dimension}
}
\value{
a \code{\link{DISTRIBUTION}}
}
\description{
This function returns a \code{\link{DISTRIBUTION}} with a new dimension
created by the sum of each row of the drawns of the distribution.
}
\details{
Only works with multidimensional distributions.
}
