% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convdistr.R
\docType{package}
\name{convdistr}
\alias{convdistr}
\title{convdistr: A package useful for convolution of distributions.}
\description{
The convdistr package provides tools to define \code{\link{DISTRIBUTION}} objects and make 
mathematical operations with them. It keeps track of the results as if they
were scalar numbers but maintaining the ability to obtain random samples of
the convoluted distributions.
}
