% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{DISTRIBUTION}
\alias{DISTRIBUTION}
\title{DISTRIBUTION class}
\description{
DISTRIBUTION is an abstract class (or interphase)
that the specific constructors should implement.
It contains 4 fields
\describe{
 \item{distribution}{A character with the name of the distribution implemented}
 \item{seed}{A numerical that is used for \code{\link{details}}} to produce
 reproducible details of the distribution
 \item{oval}{Observerd value. Is the value expected. It is used as a
 number for the mathematical operations of the distributions
 as if they were a simple scalar}
 \item{rfunc}{A function that generate random numbers from the distribution.
 Its only parameter \code{n} is the number of drawns of the distribution.
 It returns a matrix with as many rows as n, and as many columns as the
 dimensions of the distributions}
}
The DISTRIBUTION objects could support multidimensional distributions
for example  \code{\link{new_DIRICHLET}}. The names of the dimensions
should coincides with the names of the \code{oval} vector.
If only one dimension, the default name is \code{rvar}.
}
\details{
It is expected that the \code{rfunc} is included in the creation of new
distributions by convolution so the enviromnent should be carefully controlled
to avoid reference leaking that is possible within the R language. For that
reason, \code{rfunc} should be created within a \code{\link{restrict_envirnoment}}
function
}
