% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{fitbeta}
\alias{fitbeta}
\title{Fits a beta distribution based on quantiles}
\usage{
fitbeta(point, lci, uci)
}
\arguments{
\item{point}{Point estimates corresponding to the median}

\item{lci}{Lower limit (quantile 0.025)}

\item{uci}{Upper limit (quantile 0.975)}
}
\value{
parameters shape1 and shape2 of a beta distribution
}
\description{
This is a wrap of the \code{\link[SHELF]{fitdist}} to obtain
the best parameters for a beta distribution based
on quantiles.
}
\examples{
fitbeta(0.45,0.40,0.50)
}
\seealso{
\code{\link[SHELF]{fitdist}}
}
\author{
John J. Aponte
}
