% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{TRIANGULAR}
\alias{TRIANGULAR}
\alias{new_TRIANGULAR}
\title{Factory for a TRIANGULAR distribution object}
\usage{
new_TRIANGULAR(p_min, p_max, p_mode)
}
\arguments{
\item{p_min}{A numeric that represents the lower limit}

\item{p_max}{A numeric that represents the upper limit}

\item{p_mode}{A numeric that represents the mode}
}
\value{
An object of class \code{DISTRIBUTION}, \code{TRIANGULAR}
}
\description{
Returns an TRIANGULAR distribution object that produce random numbers
from a  triangular distribution using the \code{\link[extraDistr]{rtriang}}
function
}
\examples{
myDistr <- new_TRIANGULAR(-1,1,0)
myDistr$rfunc(10)
}
\author{
John Aponte
}
